+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [NEXT]
 
 - Pass `form` as second argument to `onSubmit`
+- Add `index.d.ts` types file
 
 ---